# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Lunar Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:80",
    "basePath": "/",
    "paths": {
...
        "/api/modules": {
            "get": {
                "description": "Возвращает все доступные модули с опциональной фильтрацией по названию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Модули"
                ],
                "summary": "Получить все модули",
                "parameters": [
                    {
                        "type": "string",
                        "description": "название модуля для фильтрации",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.GetAllModulesResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавить новый модуль",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Модули"
                ],
                "summary": "Добавить модуль",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Изображение модуля",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Название",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Описание",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Масса",
                        "name": "mass",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Длина",
                        "name": "length",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Диаметр",
                        "name": "diameter",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/modules/{id}": {
            "get": {
                "description": "Возвращает более подробную информацию об одном модуле",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Модули"
                ],
                "summary": "Получить один модуль",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id модуля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Module"
                        }
                    }
                }
            },
            "put": {
                "description": "Изменить данные полей о модуле",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Модули"
                ],
                "summary": "Изменить модуль",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор модуля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение модуля",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Название",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Описание",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Масса",
                        "name": "mass",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Длина",
                        "name": "length",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Диаметр",
                        "name": "diameter",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "description": "Удаляет модуль по id",
                "tags": [
                    "Модули"
                ],
                "summary": "Удалить модуль",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id модуля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/modules/{id}/add_to_mission": {
            "post": {
                "description": "Добавить выбранный модуль в черновик миссии",
                "tags": [
                    "Модули"
                ],
                "summary": "Добавить в миссию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id модуля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.Module": {
            "type": "object",
            "required": [
                "description",
                "diameter",
                "length",
                "mass"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "diameter": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "length": {
                    "type": "string"
                },
                "mass": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "schemes.GetAllModulesResponse": {
            "type": "object",
            "properties": {
                "draft_mission": {
                    "type": "string"
                },
                "modules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Module"
                    }
                }
            }
        },
...
    }
}